change set to 8bit int to save memory

+++ ext/mytest/mytest.c
@@ -21,7 +21,7 @@ long hash(char *, int, int, int);
 
 long set_len;
 int set_count;
-int *set;
+uint8_t *set;
 int kmer_size;
 int *offsets;
